Initial Commit
Basic blockchain implementation and demo of how to use it. The lib will
make a Genesis block and this block will contain random data in the
command and results. There is also a rsa key pair created to sign any
block created and the public key is exported to the previous hash
section.

Blocks can be validated by either their hash or their signature. The
hash of the block being created consists of the following and in this
order:

* Uint32 (Big Endian)
* Command
* Results of Command
* Previous hash
* Time Stamp (Big Endian)
* Signed Data

The signed data consists of the same data as above in the same order.

As of now the difficulty is based on the number of 0s at the start of a
hash. This will be modified later to be more precise.