Added a real Proof of Work (PoW)
Instead of basing the difficulty off the number of 0s in front of a hash, we now have a real PoW.
After each difficulty block is mined the target difficulty is modified based on how long it took to mine the previous blocks.
The shortened difficulty is called the target bits and stored in the block when hashed. The first byte in the target bit is the
number of 0s in the front. The rest of the target bit is the following
bytes that are not 0.